SQL refactored
